Fix dependency issue
